+++ main.go
@@ -26,7 +26,7 @@ import (
 
 // version is overridable at build time via -ldflags "-X main.version=...".
 // The Makefile injects the value from the VERSION file / git.
-var version = "0.11.1"
+var version = "0.11.2"
 
 func main() {
 	log.SetFlags(0)
@@ -64,6 +64,7 @@ func main() {
 		Version: version,
 	}, nil)
 
+	tools.ServerVersion = version
 	tools.Register(srv)
 	guides.Register(srv)